pyproject.toml
[project]
name = "scmxpert"
version = "0.1.0"
description = "SCMXpert Implementation Using CWAN"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymongo (>=4.11.3,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "passlib[bcrpt] (>=1.7.4,<2.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "datetime (>=5.5,<6.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
